Create country_redirect.js
This script will detect country location of the visitor by calling to ipinfo.io API.

In this use case, I have three server for my website, default is located in Singapore, one located in New Jersey, and the other located in Hong Kong. 
All three server has been configured with this javascript in the footer part.

This script requires jQuery to work ($.get), you can convert it to Vanilla Javascript if you have the time.